refactor: extract Report and monitor_process to _tracker.py
