+++ src/con_duct/_constants.py
@@ -1,5 +1,7 @@
 """Constants used throughout con-duct."""
 
+__schema_version__ = "0.2.2"
+
 ENV_PREFIXES = ("PBS_", "SLURM_", "OSG")
 SUFFIXES = {
     "stdout": "stdout",
