+++ test/duct_main/test_report.py
@@ -7,7 +7,8 @@ import subprocess
 from unittest import mock
 import pytest
 from con_duct._models import Averages, ProcessStats, Sample
-from con_duct.duct_main import EXECUTION_SUMMARY_FORMAT, Report
+from con_duct._tracker import Report
+from con_duct.duct_main import EXECUTION_SUMMARY_FORMAT
 
 stat0 = ProcessStats(
     pcpu=0.0,
@@ -213,7 +214,7 @@ def test_process_stats_red(
         )
 
 
-@mock.patch("con_duct.duct_main.LogPaths")
+@mock.patch("con_duct._tracker.LogPaths")
 def test_system_info_sanity(mock_log_paths: mock.MagicMock) -> None:
     mock_log_paths.prefix = "mock_prefix"
     cwd = os.getcwd()
@@ -229,9 +230,9 @@ def test_system_info_sanity(mock_log_paths: mock.MagicMock) -> None:
     assert report.system_info.user == os.environ.get("USER")
 
 
-@mock.patch("con_duct.duct_main.shutil.which")
-@mock.patch("con_duct.duct_main.subprocess.check_output")
-@mock.patch("con_duct.duct_main.LogPaths")
+@mock.patch("con_duct._tracker.shutil.which")
+@mock.patch("con_duct._tracker.subprocess.check_output")
+@mock.patch("con_duct._tracker.LogPaths")
 def test_gpu_parsing_green(
     mock_log_paths: mock.MagicMock, mock_sp: mock.MagicMock, _mock_which: mock.MagicMock
 ) -> None:
@@ -257,10 +258,10 @@ def test_gpu_parsing_green(
     ]
 
 
-@mock.patch("con_duct.duct_main.lgr")
-@mock.patch("con_duct.duct_main.shutil.which")
-@mock.patch("con_duct.duct_main.subprocess.check_output")
-@mock.patch("con_duct.duct_main.LogPaths")
+@mock.patch("con_duct._tracker.lgr")
+@mock.patch("con_duct._tracker.shutil.which")
+@mock.patch("con_duct._tracker.subprocess.check_output")
+@mock.patch("con_duct._tracker.LogPaths")
 def test_gpu_call_error(
     mock_log_paths: mock.MagicMock,
     mock_sp: mock.MagicMock,
@@ -277,10 +278,10 @@ def test_gpu_call_error(
     mlgr.warning.assert_called_once()
 
 
-@mock.patch("con_duct.duct_main.lgr")
-@mock.patch("con_duct.duct_main.shutil.which")
-@mock.patch("con_duct.duct_main.subprocess.check_output")
-@mock.patch("con_duct.duct_main.LogPaths")
+@mock.patch("con_duct._tracker.lgr")
+@mock.patch("con_duct._tracker.shutil.which")
+@mock.patch("con_duct._tracker.subprocess.check_output")
+@mock.patch("con_duct._tracker.LogPaths")
 def test_gpu_parse_error(
     mock_log_paths: mock.MagicMock,
     mock_sp: mock.MagicMock,
